refactor(pagination): no shared code
instead of implementing the code twice, wrap the headless in the daisy (not finished yet)
