+++ packages/shared/src/index.ts
@@ -1,2 +1,1 @@
 export * from './utils/clsq';
-export * from './utils/pagination';
